+++ changelog.md
@@ -1,6 +1,7 @@
 # 5.xx.xx - xx.xx.2024
 
 **What's New**
+* Remove 'prop-types' from UUI packages dependencies. Remove `uuiContextTypes` interface, since it was needed for already outdated React context API.
 * [Alert]: added support for size theming
 * [DataTable]: added support of column groups in table header. Read more - https://uui.epam.com/documents?id=advancedTables&mode=doc&category=tables&theme=electric#table_with_header_groups
 * [ErrorPage]: added field for additional message with support link (it works with `500, 503, default` errors), added support link for the site